Kept only ci_sanity tests for vdns tests
